try the fix clipppppy

+++ src/de_impl.rs
@@ -307,7 +307,7 @@ impl<'de> de::Visitor<'de> for DeserializeBoolWithVisitor {
             0 => Ok(false),
             1 => Ok(true),
             other => Err(de::Error::invalid_value(
-                Unexpected::Unsigned(other as u64),
+                Unexpected::Unsigned(other),
                 &"zero or one",
             )),
         }
